force to int on unused class
git-svn-id: http://svn.vuze.com/public/client@16150 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus3/src/com/aelitis/azureus/ui/swt/skin/SWTSkinObjectText1.java
@@ -133,9 +133,9 @@ public class SWTSkinObjectText1 extends SWTSkinObjectBasic implements
 					int iSize = NumberFormat.getInstance(Locale.US).parse(sSize).intValue();
 
 					if (firstChar == '+') {
-						iFontSize = fd[0].height + iSize;
+						iFontSize = (int)(fd[0].height + iSize);
 					} else if (firstChar == '-') {
-						iFontSize = fd[0].height - iSize;
+						iFontSize = (int)(fd[0].height - iSize);
 					} else {
 						iFontSize = iSize;
 					}
